Clean up canvas behavior
Now it passes in frame delta time instead of time elapsed since start.
This leaves it up to callers to manage the elapsed time themselves,
giving them a lot more control, e.g. adding slowdown effects or ignoring
deltas for a while if things are paused.

Also, now using the name "render" for consistency, and also min and max
limits.